asset-manager-174 Don't create unused directory
project asset-manager
taskType BUG
time 4h 21m
What was happening weas that logging was being invoked in the ArgParser before the logging directory
had been set. Parser's logging has been removed (if developers need diagnostics, they can write a query for the -d flag in the parser and use that, or query the parser in the shell after logging is configured.